Implement scaled analogue setspeed control
Since we're operating relativistically, useful speed ranges can be anywhere from 200m/s to 2000km/sec.

To make this vaugely bearable with an analogue scale type control, we actually use two controls -- a scaling slider which lets us set the magnitude of the main speed control, and the main speed control itself.

If you don't have the scale control set/bound, we default to a maximum of 200m/s, which is good enough for manuvering, but useless for interplanetary (use manual or the autopilot).